neetss-planner-improved/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
pydantic[email]==2.9.2
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
reportlab==4.2.2
bcrypt==4.0.1
